+++ experiments/libunbound.js
@@ -124,7 +124,7 @@ class LibunboundWorker {
 					if (OS) {
 						return OS.Path.join(OS.Constants.Path.profileDir, e);
 					}
-					return PathUtils.join(PathUtils.profileDir, e);
+					return PathUtils.join(PathUtils.profileDir, ...e.split(/\/|\\/));
 				}).
 				join(";");
 		} else {
@@ -401,7 +401,7 @@ this.libunbound = class extends ExtensionCommon.ExtensionAPI {
 					const res = await libunboundWorker.resolve(name, LibunboundWorker.Constants.RR_TYPE_TXT);
 					const data = res.havedata ? res.data.map(rdata => {
 						if (typeof rdata !== "string") {
-							throw Error(`DNS result has unexpected type ${typeof rdata}`);
+							throw new Error(`DNS result has unexpected type ${typeof rdata}`);
 						}
 						return rdata;
 					}) : null;