+++ test/unittest/arhParserSpec.mjs.js
@@ -151,6 +151,22 @@ describe("ARH Parser [unittest]", function () {
 			expect(res.authres_version).to.be.equal(1);
 			expect(res.resinfo.length).to.be.equal(0);
 		});
+		it("Mixed case result value for SPF (e.g. RFC 4408)", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; spf=Pass\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("spf");
+			expect(res.resinfo[0]?.result).to.be.equal("pass");
+		});
+		it("Don't restrict result keyword for unknown methods", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; unknown=foo\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("unknown");
+			expect(res.resinfo[0]?.result).to.be.equal("foo");
+		});
 	});
 	describe("Relaxed parsing", function () {
 		it("Trailing ;", function () {
@@ -166,7 +182,7 @@ describe("ARH Parser [unittest]", function () {
 			expect(res.resinfo[0]?.method).to.be.equal("spf");
 			expect(res.resinfo[0]?.result).to.be.equal("pass");
 		});
-		it("Property with / not in quotes ", function () {
+		it("Property with '/' not in quotes", function () {
 			const arh =
 				"Authentication-Results: example.com;\r\n" +
 				"          dkim=pass header.b=gfT/i2HB\r\n";
@@ -179,6 +195,45 @@ describe("ARH Parser [unittest]", function () {
 			expect(res.resinfo[0]?.method).to.be.equal("dkim");
 			expect(res.resinfo[0]?.result).to.be.equal("pass");
 			expect(res.resinfo[0]?.propertys.header.b).to.be.equal("gfT/i2HB");
+		});
+		it("Property with ':' not in quotes", function () {
+			const arh =
+				"Authentication-Results: example.com;\r\n" +
+				"  bimi=pass policy.authority-uri=\r\n" +
+				"  https://d3frv9g52qce38.cloudfront.net/amazondefault/amazon_web_services_inc.pem\r\n";
+
+			expect(() => ArhParser.parse(arh)).to.throw();
+
+			const res = ArhParser.parse(arh, true);
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("bimi");
+			expect(res.resinfo[0]?.result).to.be.equal("pass");
+			expect(res.resinfo[0]?.propertys.policy["authority-uri"]).
+				to.be.equal("https://d3frv9g52qce38.cloudfront.net/amazondefault/amazon_web_services_inc.pem");
+		});
+	});
+	describe("Invalid examples", function () {
+		it("Unknown results for DKIM", function () {
+			expect(() => ArhParser.parse(
+				"Authentication-Results: example.com; dkim=Pass\r\n"
+			)).to.throw();
+			expect(() => ArhParser.parse(
+				"Authentication-Results: example.com; dkim=hardfail\r\n"
+			)).to.throw();
+			expect(() => ArhParser.parse(
+				"Authentication-Results: example.com; dkim=foo\r\n"
+			)).to.throw();
+		});
+		it("Unknown results for SPF", function () {
+			expect(() => ArhParser.parse(
+				"Authentication-Results: example.com; spf=foo\r\n"
+			)).to.throw();
+		});
+		it("Unknown results for DMARC", function () {
+			expect(() => ArhParser.parse(
+				"Authentication-Results: example.com; dmarc=foo\r\n"
+			)).to.throw();
 		});
 	});
 	describe("DKIM results", function () {
@@ -234,6 +289,67 @@ describe("ARH Parser [unittest]", function () {
 			expect(res.resinfo[0]?.method).to.be.equal("dkim");
 			expect(res.resinfo[0]?.result).to.be.equal("fail");
 			expect(res.resinfo[0]?.reason).to.be.equal("reason quoted string");
+		});
+	});
+	describe("BIMI", function () {
+		it("B.1.  Successful BIMI lookup", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; bimi=pass header.d=example.com header.selector=myselector\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("bimi");
+			expect(res.resinfo[0]?.result).to.be.equal("pass");
+		});
+		it("B.2.  No BIMI record", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; bimi=none\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("bimi");
+			expect(res.resinfo[0]?.result).to.be.equal("none");
+		});
+		it("B.3.  Declination to Publish", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; bimi=declined\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("bimi");
+			expect(res.resinfo[0]?.result).to.be.equal("declined");
+		});
+		it("C.4.  MTA appends to Authentication-Results", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; spf=fail smtp.mailfrom=example.com;\r\n" +
+				"  dkim=pass (signature was verified) header.d=example.com;\r\n" +
+				"  dmarc=pass header.from=example.com;\r\n" +
+				"  bimi=pass header.d=example.com header.selector=brand\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(4);
+			expect(res.resinfo[3]?.method).to.be.equal("bimi");
+			expect(res.resinfo[3]?.result).to.be.equal("pass");
+		});
+		it("Result skipped", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: example.com; bimi=skipped (DMARC Policy is not at enforcement)\r\n");
+			expect(res.authserv_id).to.be.equal("example.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("bimi");
+			expect(res.resinfo[0]?.result).to.be.equal("skipped");
+		});
+		it("Example from Fastmail", function () {
+			const res = ArhParser.parse(
+				"Authentication-Results: mx1.messagingengine.com;\r\n" +
+				"    bimi=pass header.d=amazon.com header.selector=default\r\n" +
+				'      policy.authority=pass policy.mark-type="Registered Mark"\r\n' +
+				"      policy.authority-uri=\r\n" +
+				"      https://d3frv9g52qce38.cloudfront.net/amazondefault/amazon_web_services_inc.pem\r\n",
+				true);
+			expect(res.authserv_id).to.be.equal("mx1.messagingengine.com");
+			expect(res.resinfo.length).to.be.equal(1);
+			expect(res.resinfo[0]?.method).to.be.equal("bimi");
+			expect(res.resinfo[0]?.result).to.be.equal("pass");
+			expect(res.resinfo[0]?.propertys.header.selector).to.be.equal("default");
+			expect(res.resinfo[0]?.propertys.policy.authority).to.be.equal("pass");
+			expect(res.resinfo[0]?.propertys.policy["authority-uri"]).to.be.equal("https://d3frv9g52qce38.cloudfront.net/amazondefault/amazon_web_services_inc.pem");
 		});
 	});
 	describe("Internationalized Email", function () {